packages/cartesia-sdk-python/requirements.txt
# Core dependencies for supermemory-cartesia
supermemory>=3.16.0
pydantic>=2.10.0
loguru>=0.7.3
cartesia-line>=0.2.0
